odin/requirements.txt
dnspython (check name)
